+++ src/main/java/fr/estecka/invarpaint/api/PaintStackUtil.java
@@ -2,14 +2,14 @@ package fr.estecka.invarpaint.api;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import net.minecraft.component.DataComponentTypes;
-import net.minecraft.component.type.NbtComponent;
 import net.minecraft.entity.Entity;
+import net.minecraft.entity.decoration.painting.PaintingVariant;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
-import net.minecraft.nbt.NbtCompound;
+import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.util.Identifier;
-import static net.minecraft.component.DataComponentTypes.ENTITY_DATA;
+import static net.minecraft.component.DataComponentTypes.ITEM_MODEL;
+import static net.minecraft.component.DataComponentTypes.PAINTING_VARIANT;
 import static fr.estecka.invarpaint.InvarpaintMod.CONFIG;
 
 public final class PaintStackUtil
@@ -17,37 +17,53 @@ public final class PaintStackUtil
 	static public final Identifier INVALID_MODEL = Identifier.of("invarpaint", "missing_painting");
 	static public final String VARIANT_MODEL_PREFIX = "painting/";
 
-	static private final String VARIANT_TAG = "variant";
-	static private final String ENTITY_TYPE_TAG = "id";
 
+/******************************************************************************/
+/* # Set Variant                                                              */
+/******************************************************************************/
 
-	static public ItemStack	SetVariant(ItemStack stack, @NotNull Entity entity) { return SetVariant(stack, GetVariantName(entity)); }
-	static public ItemStack	SetVariant(ItemStack stack, @NotNull Identifier variantId) { return SetVariant(stack, variantId.toString()); }
-	static public ItemStack	SetVariant(ItemStack stack, @NotNull String variantName){
-		NbtCompound entityTag;
-		NbtComponent component = stack.get(ENTITY_DATA);
-		if (component != null)
-			entityTag = component.copyNbt();
-		else {
-			entityTag = new NbtCompound();
-			entityTag.putString(ENTITY_TYPE_TAG, "minecraft:painting");
-		}
-
-		entityTag.putString(VARIANT_TAG, variantName);
-		stack.set(ENTITY_DATA, NbtComponent.of(entityTag));
+	static public ItemStack	SetVariant(ItemStack stack, @NotNull RegistryEntry<PaintingVariant> entry) {
+		stack.set(PAINTING_VARIANT, entry);
 		if (CONFIG.setItemModel)
-			SetModel(stack, variantName);
+			SetModel(stack, entry);
 		return stack;
 	}
 
-	static public ItemStack	CreateVariant(Entity entity){ return CreateVariant(GetVariantName(entity)); }
-	static public ItemStack	CreateVariant(Identifier variantId){ return CreateVariant(variantId.toString()); }
-	static public ItemStack	CreateVariant(String variantName){
-		ItemStack stack = new ItemStack(Items.PAINTING);
-		SetVariant(stack, variantName);
-		if (CONFIG.setItemModel)
-			SetModel(stack, variantName);
+	static public ItemStack	SetVariant(ItemStack stack, @NotNull Entity entity) {
+		var entry = GetVariantEntry(entity);
+		if (entry != null)
+			SetVariant(stack, entry);
 		return stack;
+	}
+
+
+/******************************************************************************/
+/* # Create Variant                                                           */
+/******************************************************************************/
+
+	static public ItemStack	CreateVariant(@NotNull RegistryEntry<PaintingVariant> entry) {
+		ItemStack stack = new ItemStack(Items.PAINTING);
+		SetVariant(stack, entry);
+		if (CONFIG.setItemModel)
+			SetModel(stack, entry);
+		return stack;
+	}
+
+	static public ItemStack	CreateVariant(Entity entity){
+		var entry = GetVariantEntry(entity);
+		if (entry != null)
+			return CreateVariant(entry);
+		else
+			return new ItemStack(Items.PAINTING);
+	}
+
+
+/******************************************************************************/
+/* # Set Model                                                                */
+/******************************************************************************/
+
+	static public ItemStack	SetModel(ItemStack stack, @NotNull RegistryEntry<PaintingVariant> entry) {
+		return SetModel(stack, entry.getKey().get().getValue());
 	}
 
 	static public ItemStack SetModel(ItemStack stack, String variantName){
@@ -55,49 +71,65 @@ public final class PaintStackUtil
 		if (id != null)
 			return SetModel(stack, id);
 		else {
-			stack.set(DataComponentTypes.ITEM_MODEL, INVALID_MODEL);
+			stack.set(ITEM_MODEL, INVALID_MODEL);
 			return stack;
 		}
 	}
+
 	static public ItemStack SetModel(ItemStack stack, Identifier variantId){
-		stack.set(DataComponentTypes.ITEM_MODEL, variantId.withPrefixedPath(VARIANT_MODEL_PREFIX));
+		stack.set(ITEM_MODEL, variantId.withPrefixedPath(VARIANT_MODEL_PREFIX));
 		return stack;
 	}
 
+
+/******************************************************************************/
+/* # Get Variant                                                              */
+/******************************************************************************/
+// TODO: Probably not compatible with NoKebab
+
+	static public @Nullable RegistryEntry<PaintingVariant>	GetVariantEntry(ItemStack stack){
+		return stack.get(PAINTING_VARIANT);
+	}
+
+	static public @Nullable RegistryEntry<PaintingVariant>	GetVariantEntry(Entity entity){
+		return entity.get(PAINTING_VARIANT);
+	}
+
 	static public @Nullable Identifier	GetVariantId(ItemStack stack){
-		String name = GetVariantName(stack);
-		return (name == null) ? null : Identifier.tryParse(GetVariantName(stack));
+		var entry = GetVariantEntry(stack);
+		return (entry == null) ? null : entry.getKey().get().getValue();
 	}
 
-	/**
-	 * @implNote This particular implementation is compatible with NoKebab.
-	 */
+	@Deprecated
 	static public @Nullable String	GetVariantName(Entity entity){
-		return entity.writeNbt(new NbtCompound()).getString("variant");
+		var entry = entity.get(PAINTING_VARIANT);
+		return (entry == null) ? null : entry.getKey().get().getValue().toString();
 	}
 
+	@Deprecated
 	static public @Nullable String	GetVariantName(ItemStack stack){
-		NbtComponent entitydata = stack.get(ENTITY_DATA);
-		if (entitydata == null || !entitydata.contains(VARIANT_TAG))
-			return null;
-
-		return entitydata.copyNbt().getString(VARIANT_TAG);
+		var id = GetVariantId(stack);
+		return (id == null) ? null : id.toString();
 	}
+
+
+/******************************************************************************/
+/* # HasVariant                                                               */
+/******************************************************************************/
 
 	/**
 	 * Checks that a variant exists in any form
 	 */
 	static public boolean	HasVariant(ItemStack stack){
-		NbtComponent nbt = stack.get(ENTITY_DATA);
-		return nbt != null
-		    && nbt.contains(VARIANT_TAG)
-		    ;
+		return stack.contains(PAINTING_VARIANT);
 	}
 
 	/**
 	 * Checks that the variant is a valid identifier.
+	 * @deprecated Stacks can no longer contains invalid variants.
 	 */
+	@Deprecated
 	static public boolean	HasVariantId(ItemStack stack){
-		return GetVariantId(stack) != null;
+		return HasVariant(stack);
 	}
 }