+++ changelog.md
@@ -78,3 +78,9 @@ Removals:
 - Marked embedded datapack as compatible with MC 1.21.4
 ### 3.0.4
 - Updated for MC 1.21.5
+
+# v4
+- The resource pack no longer require the `item_model` component to be set.
+- The server no longer set the `item_model` component by default.
+- Breaking API changes, caused by minecraft internal changes.
+- Paintings with invalid variants are no longer supported by the base game. Breaks compatibility with NoKebab